\docType{data}
\name{parameter.constant}
\alias{parameter.constant}
\title{Physical constants in SI units, source: [PDG]}
\format{\preformatted{List of 8
 $ c      : num 3e+08
 $ me     : num 9.11e-31
 $ h      : num 6.63e-34
 $ mp     : num 1.67e-27
 $ e      : num 1.6e-19
 $ G      : num 6.67e-11
 $ k      : num 1.38e-23
 $ sigma.T: num 6.65e-29
}}
\usage{
parameter.constant
}
\description{
Useful physical constants in SI units.
}
\keyword{datasets}

