\docType{data}
\name{parameter.fidcosmo}
\alias{parameter.fidcosmo}
\title{Fiducial cosmology parameter list}
\format{\preformatted{List of 4
 $ omegaM0: num 0.3
 $ omegaL0: num 0.7
 $ omegaK : num 0
 $ h      : num 0.7
}}
\usage{
parameter.fidcosmo
}
\description{
the parameter list is a crucial input for other
calculations
}
\keyword{datasets}

