\docType{data}
\name{parameter.unit}
\alias{parameter.unit}
\title{useful conversion of astrophysical units}
\format{\preformatted{List of 3
 $ Mpc : num 3.09e+22
 $ Msun: num 1.99e+30
 $ eV  : num 1.6e-19
}}
\usage{
parameter.unit
}
\description{
useful conversion of astrophysical units
}
\keyword{datasets}

